Add files via upload
These files include the R program for processing and analyzing the data of mortality rates of 50 different causes for 15 countries. The program initially parses the data, adds a column with binary value to indicate if a comorbidity with COVID-19 happened for a specific year, and eventually fits a linear regression model to the data to find the correlation between each cause of death and COVID-19.